+++ src/nvim/globals.h
@@ -638,8 +638,8 @@ EXTERN int State INIT(= MODE_NORMAL);
 
 EXTERN bool debug_mode INIT(= false);
 EXTERN bool finish_op INIT(= false);    // true while an operator is pending
-EXTERN long opcount INIT(= 0);          // count for pending operator
-EXTERN int motion_force INIT(= 0);       // motion force for pending operator
+EXTERN int opcount INIT(= 0);           // count for pending operator
+EXTERN int motion_force INIT(= 0);      // motion force for pending operator
 
 // Ex Mode (Q) state
 EXTERN bool exmode_active INIT(= false);  // true if Ex mode is active
